Refactor: feature Dialog 분할.
