+++ src/entities/post/model/hooks/index.ts
@@ -1,5 +1,5 @@
-export * from './usePostsState.ts';
-export * from './useSelectionState.ts';
-export * from './useDialogState.ts';
-export * from './usePostActions.ts';
-export * from './useCommentActions.ts';
+export * from './usePostsState';
+export * from './useSelectionState';
+export * from './useDialogState';
+export * from './usePostActions';
+export * from './useCommentActions';
