+++ src/features/UserModal/index.tsx
@@ -1,10 +1,10 @@
 import * as DialogPrimitive from '@radix-ui/react-dialog';
 import React from 'react';
 
-import { useDialogState, useSelectionState } from '../../entities/post/model/hooks/index';
+import { useDialogState, useSelectionState } from '../../entities/post/model/hooks';
 import { DialogContent, DialogHeader, DialogTitle } from '../../shared/ui';
 
-export const Dialog = DialogPrimitive.Root;
+const Dialog = DialogPrimitive.Root;
 
 const UserModal: React.FC = () => {
   const { showUserModal, setShowUserModal } = useDialogState();
