+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 import './index.css';
-import App from './apps/App.tsx';
+import App from './apps/App';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
